requirements.txt
django==3.0.7
discord.py==1.3.3
python-decouple==3.3
pylint-django==2.0.15